+++ src/llama_cpp_agent/providers/kobold_cpp_endpoint_provider_old.py
