+++ src/llama_cpp_agent/providers/llama_cpp_endpoint_provider_old.py
