+++ src/llama_cpp_agent/providers/openai_endpoint_provider_old.py
